Move into Shared SqlNotificationEventArgs.cs (#1314)
